added process internal mutex
