+++ device/api/umd/device/utils/lock_manager.hpp
@@ -42,24 +42,33 @@ enum class MutexType {
 class LockManager {
 public:
     // This set of functions is used to manage mutexes which are system wide and not chip specific.
-    void initialize_mutex(MutexType mutex_type);
+    void initialize_mutex(
+        MutexType mutex_type, MutexImplementationType mutex_impl_type = MutexImplementationType::SYSTEM_WIDE);
     void clear_mutex(MutexType mutex_type);
     std::unique_lock<RobustMutex> acquire_mutex(MutexType mutex_type);
 
     // This set of functions is used to manage mutexes which are chip specific.
-    void initialize_mutex(MutexType mutex_type, int device_id, IODeviceType device_type = IODeviceType::PCIe);
+    void initialize_mutex(
+        MutexType mutex_type,
+        int device_id,
+        IODeviceType device_type = IODeviceType::PCIe,
+        MutexImplementationType mutex_impl_type = MutexImplementationType::SYSTEM_WIDE);
     void clear_mutex(MutexType mutex_type, int device_id, IODeviceType device_type = IODeviceType::PCIe);
     std::unique_lock<RobustMutex> acquire_mutex(
         MutexType mutex_type, int device_id, IODeviceType device_type = IODeviceType::PCIe);
 
     // This set of functions is used to manage mutexes which are chip specific. This variant accepts custom mutex name.
-    void initialize_mutex(std::string mutex_prefix, int device_id, IODeviceType device_type = IODeviceType::PCIe);
+    void initialize_mutex(
+        std::string mutex_prefix,
+        int device_id,
+        IODeviceType device_type = IODeviceType::PCIe,
+        MutexImplementationType mutex_impl_type = MutexImplementationType::SYSTEM_WIDE);
     void clear_mutex(std::string mutex_prefix, int device_id, IODeviceType device_type = IODeviceType::PCIe);
     std::unique_lock<RobustMutex> acquire_mutex(
         std::string mutex_prefix, int device_id, IODeviceType device_type = IODeviceType::PCIe);
 
 private:
-    void initialize_mutex_internal(const std::string& mutex_name);
+    void initialize_mutex_internal(const std::string& mutex_name, MutexImplementationType mutex_impl_type);
     void clear_mutex_internal(const std::string& mutex_name);
     std::unique_lock<RobustMutex> acquire_mutex_internal(const std::string& mutex_name);
 
@@ -72,7 +81,7 @@ private:
     // Maps from mutex name to an initialized mutex.
     // Mutex names are made from mutex type name or directly mutex name combined with device number.
     // Note that once LockManager is out of scope, all the mutexes will be cleared up automatically.
-    std::unordered_map<std::string, RobustMutex> mutexes;
+    std::unordered_map<std::string, std::unique_ptr<RobustMutex>> mutexes;
 };
 
 }  // namespace tt::umd
