+++ device/simulation/ttsim_device.cpp
@@ -61,7 +61,7 @@ TTSimDeviceInit::TTSimDeviceInit(const std::filesystem::path& simulator_director
 
 TTSimDevice::TTSimDevice(const TTSimDeviceInit& init) : Chip(init.get_soc_descriptor()) {
     log_info(tt::LogEmulationDriver, "Instantiating simulation device");
-    lock_manager.initialize_mutex(MutexType::TT_SIMULATOR);
+    lock_manager.initialize_mutex(MutexType::TT_SIMULATOR, MutexImplementationType::PROCESS_LOCAL);
     soc_descriptor_per_chip.emplace(0, init.get_soc_descriptor());
     arch_name = init.get_arch_name();
     target_devices_in_cluster = {0};
